fix(gateway): restructure code
